Refactored the entire project
Designed a easy way to create menus and route between them.

I need find a better way to handle requests.

I also refactored the Models as they were not implementing the new design pattern and the user was unable to follow through with who they created/inserted just that the query was requested.  This is no longer the issue and has been corrected.